backend/requirements.txt
alembic==1.12.1
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.3.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
deprecation==2.1.0
distro==1.9.0
ecdsa==0.19.1
fastapi==0.104.1
gotrue==1.3.1
greenlet==3.2.3
h11==0.14.0
httpcore==0.17.3
httptools==0.6.4
httpx==0.24.1
idna==3.10
iniconfig==2.1.0
Mako==1.3.10
MarkupSafe==3.0.2
openai==1.3.0
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
postgrest==0.13.2
pyasn1==0.6.1
pycparser==2.22
pydantic[email]==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
Pygments==2.19.2
pytest==8.4.1
pytest-asyncio==0.21.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.2
realtime==1.0.6
requests==2.31.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
storage3==0.6.1
StrEnum==0.4.15
supabase==2.2.0
supafunc==0.3.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.24.0
watchfiles==1.1.0
websockets==12.0
tiktoken==0.5.1
